+++ src/subcommand/run.rs
@@ -9,14 +9,19 @@ pub(crate) fn run(mut args: Args) -> Result<i32, ErrorMessage> {
     args.apply_default_target(&config, builder.kernel_root());
 
     if args.bin_name().is_none() {
-        let kernel_package = builder.kernel_package()
+        let kernel_package = builder
+            .kernel_package()
             .map_err(|key| format!("Kernel package not found it cargo metadata (`{}`)", key))?;
         let bins = kernel_package.targets.iter().filter(|t| t.kind == ["bin"]);
         let mut not_test = bins.filter(|t| !t.name.starts_with("test-"));
-        let bin_name = not_test.next().ok_or("no kernel executable found")?.name.to_owned();
+        let bin_name = not_test
+            .next()
+            .ok_or("no kernel executable found")?
+            .name
+            .to_owned();
         if not_test.count() > 0 {
             Err("Multiple kernel executables found. \
-            Please specify the binary you want to run as a `--bin` argument")?;
+                 Please specify the binary you want to run as a `--bin` argument")?;
         }
         args.set_bin_name(bin_name);
     }